Fix a bug in transaction serialization with witness data. The two-byte witness flag must be included in the witness size. This missing count was causing the txid to be incorrect, and consequently the Merkle root hash.
This change also adds the necessary logic for Merkle roots to report uniqueness of the binary tree node pairs, which is required for consensus that no such pairs are identical.

This is tested with a new integration test that downloads the first Segwit block and vaildates its Merkle root. This test uses a newly createed test::ProtocolSession which is a convenient abstraction for tests that must include p2p protocol behavior such as block download.